requirements.txt
asgiref==3.5.2
Django==4.1
django-debug-toolbar==3.5.0
django-phonenumber-field==6.3.0
phonenumberslite==8.12.54
psycopg2-binary==2.9.3
sqlparse==0.4.2
Pillow
